fix dark mode and local storage on refresh
