+++ src/contexts/HomeProvider.jsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, useEffect, useReducer  } from 'react';
 
-
+const apiKey = ''
 const HomeContext = createContext();
 
 const HomeProvider = ({ children }) => {
@@ -19,11 +19,7 @@ const HomeProvider = ({ children }) => {
             youtubeId: ""
         }
     
-        const apiKey = "8c9b44dff7454d2bb7def613b0bade75";
-        // "76c7a80de4fc4832927537ed53f92d14";
-        //  "f88e395dfddb4a21837e281aa658717c";
-        // "856ff9a8e5554f3198e5a473b5d101a8";
-        // "4defd47d816c4e5692caafff6528e6a2";
+        
        
         const [state, dispatch] = useReducer(reducer, initialState);
         const { data,
@@ -192,7 +188,7 @@ const HomeProvider = ({ children }) => {
             }
             getVideo()
         }, [foodTitle, isSelected])
-        console.log(state)                    
+        // console.log(state)                    
 
     return (
         <HomeContext.Provider
