+++ src/pages/Home.jsx
@@ -6,24 +6,7 @@ import { useDispatch } from 'react-redux';
 import { setDarkMode } from '../features/util/utilSlice';
 
 const Home = () => {
-    const dispatch = useDispatch();
-    const darkMode = JSON.parse(localStorage.getItem('darkmode'))
-    useEffect(() => {
-        dispatch(setDarkMode(darkMode))
-        // if(darkMode) {            
-        //     document.documentElement.classList.add('dark');
-        //     document.body.style.backgroundColor = "#310F01"
-        // }else {
-        //     document.documentElement.classList.remove('dark');
-        //     document.body.style.backgroundColor = "#E0DBDF"
-        // }
-        console.log(typeof darkMode, darkMode)
-
-      
-          
-        // localStorage.setItem('darkmode', darkMode)
-
-      },[darkMode, dispatch])
+ 
     return ( 
         <>  
             <Main/>
