+++ src/services/apiMealPlaner.js
@@ -22,7 +22,7 @@ export const getMealPlanWeek = async (hash, username) => {
     const startDate = `${year}-${month}-${day}`
     const response = await fetch(`https://api.spoonacular.com/mealplanner/${username}/week/${startDate}?apiKey=${apiKey}&hash=${hash}`)
     const data = await response.json();  
-    console.log(data)
+    // console.log(data)
     return data
 }
 
@@ -43,7 +43,7 @@ export const addMealPlanItem = async (username, hash, item) => {
         body: JSON.stringify(item)
     });
     const data = await response.json();
-    console.log(data);
+    // console.log(data);
     return data
 }
 
@@ -51,7 +51,7 @@ export const getOneMealTemplate = async (username, hash, id) => {
    try {
     const response = await fetch(`https://api.spoonacular.com/mealplanner/${username}/templates/120?apiKey=${apiKey}&hash=${hash}`)
     const data = await response.json();
-    console.log(data)
+    // console.log(data)
     return data
    } catch (error) {
     console.log(error)
